Created static Rubric Dashboard progress widget for mockup
